ALUMINI/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
dnspython==2.6.1
ecdsa==0.19.0
fastapi==0.114.0
filelock==3.13.4
Flask==3.0.3
fonttools==4.51.0
fsspec==2024.3.1
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
motor==3.5.1
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
paramiko==3.4.0
passlib==1.7.4
pillow==10.3.0
psutil==5.9.8
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
pygame==2.5.2
pymongo==4.8.0
PyNaCl==1.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
sympy==1.12
thop==0.1.1.post2209072238
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.1.47
urllib3==2.2.1
uvicorn==0.30.6
Werkzeug==3.0.3
